+++ packages/react-bindings/src/hooks/callUtilHooks.ts
@@ -36,6 +36,7 @@ export const useToggleCallRecording = () => {
       }
     } catch (e) {
       console.error(`Failed start recording`, e);
+      throw e;
     }
   }, [call, isCallRecordingInProgress]);